ws2_32: WS_get_local_ips: Fix ordering of IP addresses by metric if two addresses have the same metric (try 2)
The previous implementation required the metrics to be mutually inequal
(because of the "this_metric > last_metric" check). If the metrics were
not mutually inequal, it would write only one address per metric value to the list
and fill up the rest of the list with the magic loopback IP address 127.12.34.56.

Try 2 that uses qsort() as suggested by Ken Thomases.