ntdll: Try to handle write-watches while we're on the signal stack. (try 2)
Yes, this patch is really necessary. DotNet somehow thinks its a good idea
to store the IO status block in a write-watch protected memory area. If we
do not want to add wine_uninterrupted_write_memory _EVERYWHERE_ in the wine
code, we unfortunately really have to implement exception handling on the
signal stack. This is just a very basic hacky implementation of write-watch
handling, all other kind of exceptions are not handled properly yet.